Removed scrollbar and improved calendar design and layout

+++ healthcalendar/src/components/CalendarGrid.tsx
@@ -69,7 +69,8 @@ export default function CalendarGrid({
           const m = Number(d.slice(5, 7)) - 1
           const day = Number(d.slice(8, 10))
           const dateObj = new Date(y, m, day)
-          const weekday = new Intl.DateTimeFormat('en-GB', { weekday: 'short' }).format(dateObj)
+          // Full weekday name (e.g., Monday instead of Mon)
+          const weekday = new Intl.DateTimeFormat('en-GB', { weekday: 'long' }).format(dateObj)
           const dayLabel = String(dateObj.getDate()).padStart(2, '0')
           return (
             <div className={`cal-grid__day${d === todayISO ? ' cal-grid__day--today' : ''}`} key={d}>
